remove netcmd.php and port to laravel. Escape CLI and set it to stream output live.
