+++ app/Http/Controllers/Ajax/ResolutionController.php
@@ -1,0 +1,54 @@
+<?php
+/**
+ * ResolutionController.php
+ *
+ * -Description-
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @package    LibreNMS
+ * @link       http://librenms.org
+ * @copyright  2019 Tony Murray
+ * @author     Tony Murray <murraytony@gmail.com>
+ */
+
+namespace App\Http\Controllers\Ajax;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+
+class ResolutionController extends Controller
+{
+    public function set(Request $request)
+    {
+        $this->validate($request, [
+            'width' => 'required|numeric',
+            'height' => 'required|numeric'
+        ]);
+
+        // legacy session
+        session_start();
+        $_SESSION['screen_width'] = $request->width;
+        $_SESSION['screen_height'] = $request->height;
+        session_write_close();
+
+        // laravel session
+        session([
+            'screen_width' => $request->width,
+            'screen_height' => $request->height
+        ]);
+
+        return $request->width . 'x' . $request->height;
+    }
+}
